+++ boxsdk/__init__.py
@@ -2,10 +2,6 @@
 
 from __future__ import unicode_literals
 
-try:
-    from .auth.jwt_auth import JWTAuth
-except ImportError:
-    JWTAuth = None  # If extras are not installed, JWTAuth won't be available.
-from .auth.oauth2 import OAuth2
-from .client import Client
+from .auth import JWTAuth, OAuth2
+from .client import Client, DeveloperTokenClient
 from .object import *  # pylint:disable=wildcard-import,redefined-builtin